增加好友、排行榜、商店、查看玩家功能 增加获得奖励、使用通缉令界面

+++ Assets/Scripts/Data/MapInfoData.cs
@@ -5,7 +5,7 @@ public class MapInfoData {
 
     public string[] islandNames;
     public MinesData[] mines;
-    public long moneyBox;
+    public double moneyBox;
     public double producePerSecond;
     public int limit;
 }
